Start de-spaghetti process
